loplugin:redundantcast

+++ svx/source/sidebar/paragraph/ParaSpacingWindow.cxx
@@ -284,8 +284,8 @@ IMPL_LINK_NOARG(ParaLRSpacingWindow, ModifySpacingHdl, Edit&, void)
     if(pDisp)
     {
         SvxLRSpaceItem aMargin(SID_ATTR_PARA_LRSPACE);
-        aMargin.SetTextLeft((long)GetCoreValue(*m_pBeforeSpacing, m_eUnit));
-        aMargin.SetRight((long)GetCoreValue(*m_pAfterSpacing, m_eUnit));
+        aMargin.SetTextLeft(GetCoreValue(*m_pBeforeSpacing, m_eUnit));
+        aMargin.SetRight(GetCoreValue(*m_pAfterSpacing, m_eUnit));
         aMargin.SetTextFirstLineOfst((short)GetCoreValue(*m_pFLSpacing, m_eUnit));
 
         pDisp->ExecuteList(SID_ATTR_PARA_LRSPACE, SfxCallMode::RECORD, {&aMargin});
